Made some data structures to expose more functionality to Python - testing by assigning level pointer at runtime.

+++ Engine/ObjectLayer.cpp
@@ -3,7 +3,7 @@
 void CObjectLayer::OnPlay()
 {
 	// Initialising objects and stuff
-	for (int i = 0; i < m_GameObjects.size(); i++)
+	for (size_t i = 0; i < m_GameObjects.size(); i++)
 	{
 		m_GameObjects[i]->OnPlay();
 	}
@@ -11,7 +11,7 @@ void CObjectLayer::OnPlay()
 
 void CObjectLayer::Destroy()
 {
-	for (int i = 0; i < m_GameObjects.size(); i++)
+	for (size_t i = 0; i < m_GameObjects.size(); i++)
 	{
 		m_GameObjects[i]->Destroy();
 	}
@@ -20,7 +20,7 @@ void CObjectLayer::Destroy()
 
 void CObjectLayer::OnThink()
 {
-	for (int i = 0; i < m_GameObjects.size(); i++)
+	for (size_t i = 0; i < m_GameObjects.size(); i++)
 	{
 		if (!m_GameObjects[i]->OnThink()) return;
 	}
@@ -28,7 +28,7 @@ void CObjectLayer::OnThink()
 
 void CObjectLayer::Draw()
 {
-	for (int i = 0; i < m_GameObjects.size(); i++)
+	for (size_t i = 0; i < m_GameObjects.size(); i++)
 	{
 		m_GameObjects[i]->Draw();
 	}
